+++ flowman-spec/src/test/scala/com/dimajix/flowman/spec/target/TruncateTargetTest.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Kaya Kupferschmidt
+ * Copyright 2021-2022 Kaya Kupferschmidt
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -27,6 +27,7 @@ import com.dimajix.flowman.execution.ScopeContext
 import com.dimajix.flowman.execution.Session
 import com.dimajix.flowman.execution.Status
 import com.dimajix.flowman.execution.VerificationFailedException
+import com.dimajix.flowman.model.IdentifierRelationReference
 import com.dimajix.flowman.model.PartitionField
 import com.dimajix.flowman.model.Relation
 import com.dimajix.flowman.model.RelationIdentifier
@@ -61,7 +62,7 @@ class TruncateTargetTest extends AnyFlatSpec with Matchers with MockFactory with
         val targetSpec = ObjectMapper.parse[TargetSpec](spec)
         val target = targetSpec.instantiate(context).asInstanceOf[TruncateTarget]
 
-        target.relation should be (RelationIdentifier("some_relation"))
+        target.relation should be (IdentifierRelationReference(context, "some_relation"))
         target.partitions should be (Map(
             "p1" -> SingleValue("1234"),
             "p2" -> RangeValue("a", "x")
@@ -78,7 +79,7 @@ class TruncateTargetTest extends AnyFlatSpec with Matchers with MockFactory with
             .withRelations(Map("some_relation" -> relationTemplate))
             .build()
         val target = TruncateTarget(
-            Target.Properties(context),
+            context,
             RelationIdentifier("some_relation"),
             Map(
                 "p1" -> SingleValue("1234"),
@@ -88,30 +89,42 @@ class TruncateTargetTest extends AnyFlatSpec with Matchers with MockFactory with
 
         (relationTemplate.instantiate _).expects(*).returns(relation)
 
-        target.phases should be (Set(Phase.BUILD, Phase.VERIFY))
+        target.phases should be (Set(Phase.BUILD, Phase.VERIFY, Phase.TRUNCATE))
 
+        target.provides(Phase.VALIDATE) should be (Set())
+        target.provides(Phase.CREATE) should be (Set())
         (relation.provides _).expects().returns(Set(ResourceIdentifier.ofHiveTable("some_table")))
         (relation.resources _).expects(Map("p1" -> SingleValue("1234"),"p2" -> RangeValue("1", "3"))).returns(Set(
             ResourceIdentifier.ofHivePartition("some_table", Some("db"), Map("p1" -> "1234", "p2" -> "1")),
             ResourceIdentifier.ofHivePartition("some_table", Some("db"), Map("p1" -> "1234", "p2" -> "2"))
         ))
-
-        target.provides(Phase.VALIDATE) should be (Set())
-        target.provides(Phase.CREATE) should be (Set())
         target.provides(Phase.BUILD) should be (Set(
             ResourceIdentifier.ofHiveTable("some_table"),
             ResourceIdentifier.ofHivePartition("some_table", Some("db"), Map("p1" -> "1234", "p2" -> "1")),
             ResourceIdentifier.ofHivePartition("some_table", Some("db"), Map("p1" -> "1234", "p2" -> "2"))
         ))
         target.provides(Phase.VERIFY) should be (Set())
+        (relation.provides _).expects().returns(Set(ResourceIdentifier.ofHiveTable("some_table")))
+        (relation.resources _).expects(Map("p1" -> SingleValue("1234"),"p2" -> RangeValue("1", "3"))).returns(Set(
+            ResourceIdentifier.ofHivePartition("some_table", Some("db"), Map("p1" -> "1234", "p2" -> "1")),
+            ResourceIdentifier.ofHivePartition("some_table", Some("db"), Map("p1" -> "1234", "p2" -> "2"))
+        ))
+        target.provides(Phase.TRUNCATE) should be (Set(
+            ResourceIdentifier.ofHiveTable("some_table"),
+            ResourceIdentifier.ofHivePartition("some_table", Some("db"), Map("p1" -> "1234", "p2" -> "1")),
+            ResourceIdentifier.ofHivePartition("some_table", Some("db"), Map("p1" -> "1234", "p2" -> "2"))
+        ))
         target.provides(Phase.DESTROY) should be (Set())
 
-        (relation.requires _).expects().returns(Set(ResourceIdentifier.ofHiveDatabase("db")))
-        (relation.provides _).expects().returns(Set(ResourceIdentifier.ofHiveTable("some_table")))
         target.requires(Phase.VALIDATE) should be (Set())
         target.requires(Phase.CREATE) should be (Set())
+        (relation.requires _).expects().returns(Set(ResourceIdentifier.ofHiveDatabase("db")))
+        (relation.provides _).expects().returns(Set(ResourceIdentifier.ofHiveTable("some_table")))
         target.requires(Phase.BUILD) should be (Set(ResourceIdentifier.ofHiveDatabase("db"), ResourceIdentifier.ofHiveTable("some_table")))
         target.requires(Phase.VERIFY) should be (Set())
+        (relation.requires _).expects().returns(Set(ResourceIdentifier.ofHiveDatabase("db")))
+        (relation.provides _).expects().returns(Set(ResourceIdentifier.ofHiveTable("some_table")))
+        target.requires(Phase.TRUNCATE) should be (Set(ResourceIdentifier.ofHiveDatabase("db"), ResourceIdentifier.ofHiveTable("some_table")))
         target.requires(Phase.DESTROY) should be (Set())
 
         (relation.partitions _).expects().returns(Seq(PartitionField("p1", StringType), PartitionField("p2", IntegerType)))
@@ -119,6 +132,10 @@ class TruncateTargetTest extends AnyFlatSpec with Matchers with MockFactory with
         (relation.loaded _).expects(execution, Map("p1" -> SingleValue("1234"),"p2" -> SingleValue("2"))).returns(No)
         target.dirty(execution, Phase.BUILD) should be (Yes)
         target.dirty(execution, Phase.VERIFY) should be (Yes)
+        (relation.partitions _).expects().returns(Seq(PartitionField("p1", StringType), PartitionField("p2", IntegerType)))
+        (relation.loaded _).expects(execution, Map("p1" -> SingleValue("1234"),"p2" -> SingleValue("1"))).returns(Yes)
+        (relation.loaded _).expects(execution, Map("p1" -> SingleValue("1234"),"p2" -> SingleValue("2"))).returns(No)
+        target.dirty(execution, Phase.TRUNCATE) should be (Yes)
 
         (relation.partitions _).expects().returns(Seq(PartitionField("p1", StringType), PartitionField("p2", IntegerType)))
         (relation.loaded _).expects(execution, Map("p1" -> SingleValue("1234"),"p2" -> SingleValue("1"))).returns(No)
@@ -149,34 +166,41 @@ class TruncateTargetTest extends AnyFlatSpec with Matchers with MockFactory with
             .withRelations(Map("some_relation" -> relationTemplate))
             .build()
         val target = TruncateTarget(
-            Target.Properties(context),
+            context,
             RelationIdentifier("some_relation")
         )
 
         (relationTemplate.instantiate _).expects(*).returns(relation)
 
-        target.phases should be (Set(Phase.BUILD, Phase.VERIFY))
-
-        (relation.provides _).expects().returns(Set(ResourceIdentifier.ofHiveTable("some_table")))
-        (relation.resources _).expects(Map.empty[String,FieldValue]).returns(Set(ResourceIdentifier.ofHiveTable("some_table")))
+        target.phases should be (Set(Phase.BUILD, Phase.VERIFY, Phase.TRUNCATE))
 
         target.provides(Phase.VALIDATE) should be (Set())
         target.provides(Phase.CREATE) should be (Set())
+        (relation.provides _).expects().returns(Set(ResourceIdentifier.ofHiveTable("some_table")))
+        (relation.resources _).expects(Map.empty[String,FieldValue]).returns(Set(ResourceIdentifier.ofHiveTable("some_table")))
         target.provides(Phase.BUILD) should be (Set(ResourceIdentifier.ofHiveTable("some_table")))
         target.provides(Phase.VERIFY) should be (Set())
+        (relation.provides _).expects().returns(Set(ResourceIdentifier.ofHiveTable("some_table")))
+        (relation.resources _).expects(Map.empty[String,FieldValue]).returns(Set(ResourceIdentifier.ofHiveTable("some_table")))
+        target.provides(Phase.TRUNCATE) should be (Set(ResourceIdentifier.ofHiveTable("some_table")))
         target.provides(Phase.DESTROY) should be (Set())
 
-        (relation.requires _).expects().returns(Set(ResourceIdentifier.ofHiveDatabase("db")))
-        (relation.provides _).expects().returns(Set(ResourceIdentifier.ofHiveTable("some_table")))
         target.requires(Phase.VALIDATE) should be (Set())
         target.requires(Phase.CREATE) should be (Set())
+        (relation.requires _).expects().returns(Set(ResourceIdentifier.ofHiveDatabase("db")))
+        (relation.provides _).expects().returns(Set(ResourceIdentifier.ofHiveTable("some_table")))
         target.requires(Phase.BUILD) should be (Set(ResourceIdentifier.ofHiveDatabase("db"), ResourceIdentifier.ofHiveTable("some_table")))
         target.requires(Phase.VERIFY) should be (Set())
+        (relation.requires _).expects().returns(Set(ResourceIdentifier.ofHiveDatabase("db")))
+        (relation.provides _).expects().returns(Set(ResourceIdentifier.ofHiveTable("some_table")))
+        target.requires(Phase.TRUNCATE) should be (Set(ResourceIdentifier.ofHiveDatabase("db"), ResourceIdentifier.ofHiveTable("some_table")))
         target.requires(Phase.DESTROY) should be (Set())
 
         (relation.loaded _).expects(execution, Map.empty[String,SingleValue]).returns(Yes)
         target.dirty(execution, Phase.BUILD) should be (Yes)
         target.dirty(execution, Phase.VERIFY) should be (Yes)
+        (relation.loaded _).expects(execution, Map.empty[String,SingleValue]).returns(Yes)
+        target.dirty(execution, Phase.TRUNCATE) should be (Yes)
 
         (relation.loaded _).expects(execution, Map.empty[String,SingleValue]).returns(Yes)
         target.execute(execution, Phase.VERIFY).exception.get shouldBe a[VerificationFailedException]
@@ -189,5 +213,8 @@ class TruncateTargetTest extends AnyFlatSpec with Matchers with MockFactory with
 
         (relation.loaded _).expects(execution, Map.empty[String,SingleValue]).returns(No)
         target.dirty(execution, Phase.BUILD) should be (No)
+
+        (relation.loaded _).expects(execution, Map.empty[String,SingleValue]).returns(No)
+        target.dirty(execution, Phase.TRUNCATE) should be (No)
     }
 }